requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
av==15.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
ctranslate2==4.4.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
faster-whisper==1.2.0
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
huggingface-hub==0.35.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lightning==2.5.5
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
nltk==3.9.1
numpy==2.3.3
omegaconf==2.3.0
onnxruntime==1.22.1
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
primePy==1.3
propcache==0.3.2
protobuf==6.32.1
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.4
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.43
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.56.2
typer==0.17.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whisperx @ git+https://github.com/m-bain/whisperx.git@2d9ce44329ae73af2520196d31cd14b6192ace44
yarl==1.20.1
